Dhruva| Extracted the export to excel function
